Bluetooth: OPP: Handle Try Again exception to retry write
When the write to socket is faster than BT maximum throughput
socket buffer gets full and it returns EAGAIN, writing application
has to retry in this case. This return value is thrown as exception
from JNI layer and has to be handled properly to retry the write
operation.

CRs-fixed: 449121
Change-Id: Ie27423744e0c0974cf75352b69a66c4cd7116914

Conflicts:
	src/com/android/bluetooth/opp/BluetoothOppObexClientSession.java